+++ cmd/list.go
@@ -2,18 +2,21 @@ package cmd
 
 import (
 	"fmt"
+	"io/ioutil"
 	"log"
 	"os"
 	"text/tabwriter"
 
-	"github.com/danhale-git/craft/internal/docker"
+	"github.com/danhale-git/craft/internal/backup"
 
-	"github.com/danhale-git/craft/internal/craft"
+	"github.com/danhale-git/craft/internal/docker"
 
 	"github.com/spf13/cobra"
 )
 
-var timeFormat = "02 Jan 2006 3:04PM"
+const (
+	timeFormat = "02 Jan 2006 3:04PM"
+)
 
 func init() {
 	// listCmd represents the list command
@@ -59,7 +62,7 @@ func init() {
 			}
 
 			// List backed up servers
-			backupNames, err := craft.BackupServerNames()
+			backupNames, err := backupServerNames()
 			if err != nil {
 				log.Fatalf("Error getting backups: %s", err)
 			}
@@ -77,7 +80,7 @@ func init() {
 					continue
 				}
 
-				_, t, err := craft.LatestServerBackup(n)
+				_, t, err := backup.LatestFile(n)
 				if err != nil {
 					log.Fatalf("Error getting latest backup: %s", err)
 				}
@@ -96,4 +99,21 @@ func init() {
 	listCmd.Flags().BoolP("all", "a", false, "Show all servers. Defaults to only running servers.")
 
 	rootCmd.AddCommand(listCmd)
+}
+
+// backupServerNames returns a slice with the names of all backed up servers.
+func backupServerNames() ([]string, error) {
+	backupDir := backup.Directory()
+	infos, err := ioutil.ReadDir(backupDir)
+
+	if err != nil {
+		return nil, fmt.Errorf("reading directory '%s': %s", backupDir, err)
+	}
+
+	names := make([]string, len(infos))
+	for i, f := range infos {
+		names[i] = f.Name()
+	}
+
+	return names, nil
 }
